cmd/trayscale: move `peerPage`-related stuff into its own file
